requirement.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2
djangorestframework==3.16.1
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.3.0
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.5
django-cors-headers==4.7.0
django-redis==6.0.0
pyotp==2.9.0
whitenoise==6.9.0
gunicorn==23.0.0
